+++ sns_app/package.json
@@ -27,6 +27,7 @@
     "@nestjs/serve-static": "^4.0.0",
     "@prisma/client": "^5.0.0",
     "bcrypt": "^5.1.0",
+    "class-validator": "^0.14.0",
     "ejs": "^3.1.9",
     "express": "^4.18.2",
     "hbs": "^4.2.0",
