+++ sns_app/src/app.module.ts
@@ -3,13 +3,14 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ServeStaticModule } from '@nestjs/serve-static';
 import { join } from 'path';
-import { SignupController } from './Controller/signup/signup.controller';
+import { SignupController } from './account/signup/signup.controller';
+import { AccountService } from './Service/account/account.service';
 
 @Module({
   imports: [
     ServeStaticModule.forRoot({ rootPath: join(__dirname, '..', 'public') }),
   ],
   controllers: [AppController, SignupController],
-  providers: [AppService],
+  providers: [AppService, AccountService],
 })
 export class AppModule {}
